refactors in signatures, int to long on ID fields.

+++ src/com/sirma/internal/staffmanagementsystem/management/model/StaffService.java
@@ -30,7 +30,7 @@ public class StaffService implements Service {
     }
 
     @Override
-    public Employee search(int id) {
+    public Employee search(long id) {
         return null;
     }
